move config to bin
