+++ gpkg_bin/src/commands/list.rs
@@ -1,6 +1,6 @@
 use super::Command;
+use crate::config::Config;
 use colored::*;
-use gpkg_lib::config::Config;
 use gpkg_lib::storage::Metadata;
 
 use structopt::StructOpt;
@@ -12,7 +12,7 @@ impl Command for List {
     type Error = ();
 
     fn apply(self, config: Config) -> Result<(), Self::Error> {
-        let binaries = Metadata::read_all(&config).expect("Can't read files");
+        let binaries = Metadata::read_all(config.bin_dir()).expect("Can't read files");
         let max_width = binaries.iter().map(|x| x.binary_name.len()).max();
 
         if let Some(max_width) = max_width {
