+++ gpkg_bin/src/commands/uninstall.rs
@@ -1,6 +1,6 @@
 use super::Command;
+use crate::config::Config;
 use colored::*;
-use gpkg_lib::config::Config;
 use gpkg_lib::node_package_version::NodePackageVersion;
 use gpkg_lib::storage::Metadata;
 use structopt::StructOpt;
@@ -13,7 +13,7 @@ pub struct Uninstall {
 impl Command for Uninstall {
     type Error = ();
     fn apply(self, config: Config) -> Result<(), Self::Error> {
-        let binaries = Metadata::read_all(&config).expect("Can't read metadata files");
+        let binaries = Metadata::read_all(config.bin_dir()).expect("Can't read metadata files");
         let binaries = binaries
             .iter()
             .filter(|metadata| metadata.package_name == self.version.name());
@@ -21,12 +21,9 @@ impl Command for Uninstall {
         for binary_metadata in binaries {
             let binary_name = &binary_metadata.binary_name;
             let binary_path = config.bin_dir().join(binary_name);
-            let metadata_path = Metadata::path_for(binary_name, &config);
 
             std::fs::remove_file(&binary_path)
                 .expect(&format!("Can't delete file {:?}", binary_path));
-            std::fs::remove_file(&metadata_path)
-                .expect(&format!("Can't delete file {:?}", metadata_path));
 
             println!("Deleted binary {}", binary_metadata.binary_name.cyan());
         }
