+++ gpkg_lib/src/lib.rs
@@ -1,8 +1,8 @@
-pub mod config;
+mod binary;
 pub mod directory_portal;
 pub mod from;
 pub mod install_package;
 pub mod node_package_version;
-pub mod npm;
-pub mod package_json;
+mod npm;
+mod package_json;
 pub mod storage;
